refactor: #1 DIContainer 클래스 -> core 패키지 이동
