+++ backend/src/main/java/com/example/matcha/framework/di/core/DIContainer.java
@@ -1,7 +1,6 @@
-package com.example.matcha.framework.di;
+package com.example.matcha.framework.di.core;
 
 import java.lang.reflect.Constructor;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
