+++ backend/src/test/java/com/example/matcha/framework/di/core/DIContainerTest.java
@@ -1,4 +1,4 @@
-package com.example.matcha.framework.di;
+package com.example.matcha.framework.di.core;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertSame;
